[update] changed the overall tsx and css file structure
