+++ index.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/png" href="./src/assets/unimatelogo.png" />
+    <link rel="icon" type="image/png" href="./src/assets/logo/unimatelogo.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>UniMate</title>
   </head>
