+++ src/App.tsx
@@ -1,9 +1,9 @@
 import './App.css'
 import './services/firebase.tsx'
-import Navbar from './components/navbar.tsx';
+import Navbar from './components/layout/navbar.tsx';
 import {Route, Routes} from "react-router-dom"
 import {Mentors} from "./pages/Mentors.tsx"
-import MentorProfile from './pages/mentorProfile.tsx';
+import MentorProfile from './pages/MentorProfile.tsx';
 
 
 function App() {
