+++ src/assets/logo/unimate_logo.png
