+++ src/pages/MentorProfile.tsx
@@ -1,0 +1,10 @@
+import Overview from "../components/profileMentor/overview";
+function MentorProfile(){
+    return(
+        <>
+        <Overview/>
+        </>
+    );
+}
+
+export default MentorProfile;
