+++ src/styles/review.css
@@ -1,104 +1,3 @@
-/* Too many classes =(((((*/
-
-.emptySpace{
-    padding-bottom: 5%;
-}
-
-.main{
-    background-color: white;
-    text-align: left;
-    padding-bottom: 100%; 
-    color: black;
-}
-
-.profile{
-    max-width: 92%;
-    margin-left: 4%;
-    margin-right: 4%;
-}
-
-.generalInfo{
-    padding-bottom: 5%;
-    display: inline-block;
-}
-
-.pfp img{
-    max-width: 30%;
-    position:relative;
-    float: left;
-    padding-right: 7%;
-}
-
-.attributes{
-    line-height: 130%;
-    width: 30%;
-    position: relative;
-    float: left;
-    color: black;
-}
-
-.attributes .name{
-    font-weight: bold;
-    font-size: 150%;
-}
-
-.education{
-    padding-top: 12%;
-}
-
-.education .logo img{
-    position:relative;
-    float: left;
-    padding-right: 7%;
-}
-
-.education .eduinfo{
-    position: relative;
-    float: left;
-    font-size: small;
-}
-
-.education .eduinfo .school{
-    font-weight: bold;
-}
-
-.profileOverview{
-    padding-bottom: 5%;
-}
-.profileOverview .title{
-    color: #00239C;
-    font-weight: bold;   
-    font-size: 20px;
-    border-bottom:1px solid #00B1D2;
-    padding-bottom:1px;
-}
-
-.description{
-    max-width: 50%;
-    display: inline-block;
-    position:relative;
-    text-wrap: left;
-}
-
-.topics{
-    width: 30%;
-    position: relative;
-    float: right;
-}
-
-/*reviews section*/
-.reviews{
-    padding-bottom: 5%;
-}
-
-.reviews .title{
-    color: #00239C;
-    font-weight: bold;   
-    font-size: 20px;
-    border-bottom:1px solid #00B1D2;
-    padding-bottom:1px;
-}
-
 .reviews .reviews-filter{
     position: relative;
     display: block;
